Updated some of the maths classes and included my performance test, mat4 and quaternions classes from my GCP assignment 1.

+++ LevelHEngine/Maths/Vec2.h
@@ -65,7 +65,7 @@ namespace Maths
 		*/
 		float getLength()
 		{
-			return sqrt((x*x) + (y*y));
+			return float(sqrt((x*x) + (y*y)));
 		}
 	};
